backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
litellm==1.17.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
